cambio de long a in en el id de user

+++ Models/Image.cs
@@ -10,7 +10,6 @@ namespace Web_Service_.Net_Core.Models
     public class Image
     {
         public long Id { get; set; }
-
         public string DeleteHash { get; set; } = "";
         public string Name { get; set; } = "";
         public string Url { get; set; } = "";
@@ -20,7 +19,7 @@ namespace Web_Service_.Net_Core.Models
         [JsonIgnore]
         public long? IdProduct { get; set; }
         [JsonIgnore]
-        public long? IdUser { get; set; }
+        public int? IdUser { get; set; }
         [JsonIgnore]
         [ForeignKey("IdUser")]
         public User? User { get; set; }
